requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
tqdm==4.65.0
faiss-cpu==1.7.4
click==8.1.3
cryptography==40.0.2
pydantic==1.10.7
joblib==1.2.0

# Core security dependencies
liboqs-python==0.7.2  # Post-quantum cryptography library
shamir-mnemonic==0.2.0  # For Shamir's Secret Sharing
cryptography==40.0.2  # Modern cryptography library
pyca/cryptography==40.0.2  # For password-based encryption

# API and Web UI dependencies
flask==2.3.2
flask-restful==0.3.10
flask-jwt-extended==4.5.2
flask-cors==4.0.0
streamlit==1.22.0
streamlit-option-menu==0.3.6
fastapi==0.95.1
uvicorn==0.22.0
gunicorn==20.1.0

# Database and storage
redis==4.5.5
pymongo==4.3.3
sqlalchemy==2.0.12
alembic==1.10.4

# Enterprise features
python-ldap==3.4.3  # LDAP integration
psutil==5.9.5  # System monitoring
prometheus-client==0.16.0  # Metrics
pyyaml==6.0  # Configuration
apscheduler==3.10.1  # Scheduling for key rotation
python-dotenv==1.0.0  # Environment configuration
passlib==1.7.4  # Password policy enforcement
argon2-cffi==21.3.0  # Advanced password hashing

# High availability and clustering
kazoo==2.9.0  # ZooKeeper client for coordination
watchdog==3.0.0  # File watching for config changes

# Logging and monitoring
structlog==23.1.0  # Structured logging
python-json-logger==2.0.7  # JSON formatted logs
elasticsearch==8.7.0  # Log aggregation
opencensus==0.11.2  # Distributed tracing
sentry-sdk==1.21.1  # Error tracking

# Testing and quality
pytest==7.3.1
black==23.3.0
mypy==1.3.0
bandit==1.7.5
coverage==7.2.5

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.0
mkdocs==1.4.3
mkdocs-material==9.1.8

# Optional hardware security module support
python-pkcs11==0.7.0  # PKCS#11 library for HSM integration 